// src/generated/client/models/Trailer.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Vehicle Management Services (management)
 * Vehicle Management Services (management)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * Represent single trailer
 * @export
 * @interface Trailer
 */
export interface Trailer {
    /**
     * 
     * @type {string}
     * @memberof Trailer
     */
    readonly id?: string;
    /**
     * 
     * @type {string}
     * @memberof Trailer
     */
    plateNumber: string;
}

/**
 * Check if a given object implements the Trailer interface.
 */
export function instanceOfTrailer(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "plateNumber" in value;

    return isInstance;
}

export function TrailerFromJSON(json: any): Trailer {
    return TrailerFromJSONTyped(json, false);
}

export function TrailerFromJSONTyped(json: any, ignoreDiscriminator: boolean): Trailer {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': !exists(json, 'id') ? undefined : json['id'],
        'plateNumber': json['plateNumber'],
    };
}

export function TrailerToJSON(value?: Trailer | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'plateNumber': value.plateNumber,
    };
}